requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt[crypto]>=5.3
psycopg[binary]>=3.2
dj-database-url>=2.1
python-dotenv>=1.0
django-cors-headers>=4.4
drf-spectacular>=0.27